requirements.txt

chardet
langfuse
# Main
llama-index
llama-index-embeddings-openai
llama-index-llms-openai
# Reader/ Parser
llama-index-readers-file
llmsherpa
pre-commit
pydantic
pydantic-settings
python-dotenv
# Others
tqdm
# Others
typer
